+++ resdk/resources/__init__.py
@@ -42,6 +42,7 @@ Utility functions
 
 from .collection import Collection
 from .data import Data
+from .descriptor import DescriptorSchema
 from .process import Process
 from .relation import Relation
 from .sample import Sample